+++ docs/content/foundations/icons.mdx
@@ -13,3 +13,7 @@ Each icon in our library is meaningful and universal, to ensure users understand
 ## React component
 
 The [Icon component](/components/icon) is used to apply our set of universal icons to more complex components of the system.
+
+## All icons
+
+<AllIconsPlayground />
